Rename DeathEater to IgnoreThrowablesWrapper & reorganize namespace
